+++ jvm/src/main/scala/com/nawforce/apexlink/org/HoverProvider.scala
@@ -3,6 +3,7 @@
  */
 package com.nawforce.apexlink.org
 
+import com.nawforce.apexlink.cst.{Creator, NewExpression}
 import com.nawforce.apexlink.rpc.HoverItem
 import com.nawforce.apexlink.types.apex.{
   ApexClassDeclaration,
@@ -10,6 +11,7 @@ import com.nawforce.apexlink.types.apex.{
   ApexFullDeclaration,
   ApexMethodLike
 }
+import com.nawforce.apexlink.types.core.TypeId
 import com.nawforce.pkgforce.path.{Locatable, Location, PathLike}
 
 trait HoverProvider extends SourceOps {
@@ -30,18 +32,53 @@ trait HoverProvider extends SourceOps {
   ): Option[(Locatable, Location)] = {
     val validation = locateFromValidation(td, line, offset)
 
-    validation._2.flatMap(loc => {
-      validation._1(loc).result.locatable match {
-        case Some(l: ApexMethodLike) =>
-          Some(l, loc)
-        case Some(l: ApexConstructorLike) =>
-          Some(l, loc)
-        case Some(l: ApexClassDeclaration) =>
-          Some(l, loc)
-        case _ =>
-          None
+    // Check all validation results to see if we have a class directly
+    val allLocatables = validation._1
+      .filter { case (loc, _) => loc.contains(line, offset) }
+      .flatMap { case (loc, vr) =>
+        vr.result.locatable.map((_, loc))
       }
-    })
+      .toSeq
+
+    val classLocatable = allLocatables.collectFirst { case (l: ApexClassDeclaration, loc) =>
+      (l, loc)
+    }
+    val constructorLocatable = allLocatables.collectFirst { case (l: ApexConstructorLike, loc) =>
+      (l, loc)
+    }
+    val methodLocatable = allLocatables.collectFirst { case (l: ApexMethodLike, loc) =>
+      (l, loc)
+    }
+
+    // If we found a class directly, use it
+    // Otherwise, if we found a constructor, check if it's parameterless
+    // Parameterless constructors should show the class, constructors with parameters should show the constructor
+    methodLocatable
+      .orElse(classLocatable)
+      .orElse(constructorLocatable.flatMap { case (ctor, loc) =>
+        if (ctor.parameters.isEmpty) {
+          // Parameterless constructor - show the class
+          // For parameterless constructors, show the class declaration
+          // The location should be where the class name appears in the source (from validation result)
+          ctor.thisTypeId.toTypeDeclaration[ApexClassDeclaration] match {
+            case Some(classDecl) =>
+              // Use the validation result location start, but adjust end to match class name length
+              // The location should span just the class name, not the entire "new Dummy()" expression
+              val classNameLength = classDecl.name.value.length
+              val adjustedLocation = Location(
+                loc.startLine,
+                loc.startPosition,
+                loc.endLine,
+                loc.startPosition + classNameLength
+              )
+              Some(classDecl, adjustedLocation)
+            case None => Some(ctor, loc)
+          }
+        } else {
+          // Constructor with parameters - show the constructor
+          Some(ctor, loc)
+        }
+      })
   }
 
   private def toHoverItem(l: Option[(Locatable, Location)]): HoverItem = {